Remove stub functions related to offload playback
Following functions must be not defined if we do not support
corresponding functionality:
* pause
* resume
* drain

Upper level functions supportsPauseAndResume() and supportsDrain()
report functionality as supported if corresponding field is not NULL.
That's why we need to remove stubs for mentioned functions.
Also flush() function is removed because it depends on pause()
and also is optional.

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>